pas de paiement exeptionelle quand production
Former-commit-id: 18aa87afa9ed883f7b6eecc4b12a3476d5ea37b0

+++ htdocs/bimpcore/updates/10.5.sql
@@ -1,1 +1,1 @@
-ALTER TABLE `llx_bf_refinanceur` ADD `quantity` INT NOT NULL AFTER `date_update`, ADD `amount_ht` DECIMAL NOT NULL AFTER `quantity`, ADD `payment` INT NOT NULL AFTER `amount_ht`, ADD `periodicity` INT NOT NULL AFTER `payment`, ADD `periode2` BOOLEAN NOT NULL AFTER `periodicity`;
+ALTER TABLE `llx_bf_refinanceur` ADD `quantity` INT NOT NULL AFTER `date_update`, ADD `amount_ht` DECIMAL (12,4) NOT NULL AFTER `quantity`, ADD `payment` INT NOT NULL AFTER `amount_ht`, ADD `periodicity` INT NOT NULL AFTER `payment`, ADD `periode2` BOOLEAN NOT NULL AFTER `periodicity`;
